refactor: 뷰모델 통합
