+++ app/src/main/java/umc/OnAirMate/ui/chat_room/summary/SummaryFragment.kt
@@ -15,7 +15,7 @@ class SummaryFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         // Inflate the layout for this fragment
         binding = FragmentSummaryBinding.inflate(layoutInflater)
